Make the `i18n` tag to work
We had to make these changes for to et the `i18n` tag to work.

With the ifs around the `<script>` tags we got "unclosed script tag" as a runtime error.

Without the `.raw()` we found the double quotes in the bit of JS were all escaped (i.e. `\"` instead of `"`).

No test case, sorry.