Improve log4j.properties file loading - add failing test
In Play! 1.5.x documentation [1] the property "application.log.path"
description is the following:
Path to a Log4J configuration file, to customise log output.
If you do not specify a path, Play will load a log4j.properties file
in the conf directory if present.

The same can be found in the annotated application.conf:
"If you want a very customized log, create a log4j.properties file
in the conf directory"

I'm unsure if it worked at all in Play!, but it's a nice feature.

The init_with_conf_dir() test tries to check the parameters of the newly
added "conf/log4j.properties".

The init_with_default_config() test deletes that properties file
from the test classpath and checks for the Framework's properties file.

[1]: https://www.playframework.com/documentation/1.5.x/configuration#application.log.path