Improve log4j.properties file loading - load from conf/ automatically
In Play! 1.5.x documentation [1] the property "application.log.path"
description is the following:
Path to a Log4J configuration file, to customise log output.
If you do not specify a path, Play will load a log4j.properties file
in the conf directory if present.

The same can be found in the annotated application.conf:
"If you want a very customized log, create a log4j.properties file
in the conf directory"

I'm unsure if it worked at all in Play!, but it's a nice feature.

[1]: https://www.playframework.com/documentation/1.5.x/configuration#application.log.path