gradle: explicitly require specific version of Java toolchain
to avoid build time errors:

> Task :framework:compileJava FAILED

FAILURE: Build failed with an exception.

* What went wrong:
Execution failed for task ':framework:compileJava'.
> invalid source release: 11

Since Gradle 8.0.2 one needs to opt in for automatic toolchain
provisioning with adding plugins to "settings.gradle".

See [1] for details!

The best way to target a Java version is to build with that specific
version of Java.
If that's not possible, one could use the "release" option,
available from Java 10 within Gradle.

See [2] for details.

While at it, let Gradle configure the Kotlin test app automatically.

See [3] for details.

[1]: https://docs.gradle.org/8.5/userguide/toolchains.html
[2]: https://docs.gradle.org/8.5/userguide/building_java_projects.html#sec:java_cross_compilation
[3]: https://kotlinlang.org/docs/gradle-configure-project.html#gradle-java-toolchains-support