#327 fix CVE-2023-24998 in commons-fileupload
Apache Commons FileUpload prior to 1.5.0.redhat-00001 does not limit the number of request parts to be processed, resulting in the possibility of an attacker triggering a Denial of Service (DoS) with a malicious upload or series of uploads. Note that, like all the file upload limits, the new configuration option (FileUploadBase#setFileCountMax) is not enabled by default and must be explicitly configured.

See https://devhub.checkmarx.com/cve-details/CVE-2023-24998/