framework: fix Router.reverseWithCheck() calls with classpath files
Fix Router.reverseWithCheck() calls for classpath resources
with the help of ServerHelper.findFile() and with adding a custom route.

While at it, improve ServerHelper.findFile() to reuse the extracted
classpath resource files.

To prevent unwanted leakage of classpath files, the custom route exposes
only classpath files with path starting with "public/".
This can be improved later.